ai-service/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_client]
faster-whisper
webrtcvad
opuslib
numpy
requests
